+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Changelog
 
+* make `Fill` discriminated by its type (BREAKING: removes `FillType` enum)
+
 ## v2.0.0
 * Global type definitions removed, the usage of global types is an anti pattern