+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * make `Fill` discriminated by its type (BREAKING: removes `FillType` enum)
 * fix callback error type
+* add missing `null` types in return or argument positions
 
 ## v2.0.0
 * Global type definitions removed, the usage of global types is an anti pattern